deck: Move Card functionality into seperate mod
